+++ zpa/services/applicationsegment/zpa_application_segment_test.go
@@ -47,6 +47,7 @@ func TestApplicationSegment(t *testing.T) {
 	}()
 
 	service := New(client)
+
 	appSegment := ApplicationSegmentResource{
 		Name:                  name,
 		Description:           "New application segment",
@@ -172,7 +173,7 @@ func TestRetrieveNonExistentResource(t *testing.T) {
 	}
 	service := New(client)
 
-	_, _, err = service.Get("non_existent_id")
+	_, _, err = service.Get("non-existent-id")
 	if err == nil {
 		t.Error("Expected error retrieving non-existent resource, but got nil")
 	}
@@ -185,7 +186,7 @@ func TestDeleteNonExistentResource(t *testing.T) {
 	}
 	service := New(client)
 
-	_, err = service.Delete("non_existent_id")
+	_, err = service.Delete("non-existent-id")
 	if err == nil {
 		t.Error("Expected error deleting non-existent resource, but got nil")
 	}
@@ -198,7 +199,7 @@ func TestUpdateNonExistentResource(t *testing.T) {
 	}
 	service := New(client)
 
-	_, err = service.Update("non_existent_id", ApplicationSegmentResource{})
+	_, err = service.Update("non-existent-id", ApplicationSegmentResource{})
 	if err == nil {
 		t.Error("Expected error updating non-existent resource, but got nil")
 	}
@@ -211,7 +212,7 @@ func TestGetByNameNonExistentResource(t *testing.T) {
 	}
 	service := New(client)
 
-	_, _, err = service.GetByName("non_existent_name")
+	_, _, err = service.GetByName("non-existent-name")
 	if err == nil {
 		t.Error("Expected error retrieving resource by non-existent name, but got nil")
 	}